changing <cfml> to (cfml)
Great job on the book, Luis! That said, I notice some inconsistency where sometimes "ColdFusion <CFML>" is used, but other times "ColdFusion (CFML)". Of course, the latter may be preferred by many, as the former can trigger those with an aversion to tags, or with an aversion to CFML being associated negatively for being tag-based. And given how script is indeed the focus of most of this resource, it does seem better to avoid the brackets.

So I've changed that here. I notice it's already been done in the title in the top left corner of the online book (though the cover of the printed book uses it, and I realize that could be very hard for you to change--though it may be worth pursuing.)